Error??? ????RSS?????

+++ src/lab/cb/scmd/web/servlet/RSSErrorServlet.java
@@ -40,8 +40,8 @@ public class RSSErrorServlet extends HttpServlet {
 		channel.setChannelUri(request.getRequestURL().toString());
 		channel.setChannelLanguage("ja");
 
-		TreeMap<Integer,RssChannelItem> items = ErrorLog.getRssChannelItems();
-		for(Integer key : items.keySet()) {
+		TreeMap<Long,RssChannelItem> items = ErrorLog.getRssChannelItems();
+		for(Long key : items.keySet()) {
 			channel.addItem(items.get(key));
 		}
 		rssdoc.addChannel(channel);
